Extended the packet definitions and their usage (Server -> Client)
This also required to support struct "arrays", which are implemented by using indexers.
Extended the length calculation by considering struct arrays, too.

+++ src/GameServer/RemoteView/Guild/AssignPlayersToGuildPlugIn.cs
@@ -77,7 +77,7 @@ namespace MUnique.OpenMU.GameServer.RemoteView.Guild
         private void SetGuildPlayerBlock(Span<byte> playerBlock, Player guildPlayer, bool appearsNew)
         {
             playerBlock.SetIntegerBigEndian(guildPlayer.GuildStatus.GuildId);
-            playerBlock[4] = guildPlayer.GuildStatus.Position.GetViewValue();
+            playerBlock[4] = (byte)guildPlayer.GuildStatus.Position.GetViewValue();
 
             var playerId = guildPlayer.GetId(this.player);
             playerBlock[7] = (byte)(playerId.GetHighByte() | (appearsNew ? 0x80 : 0));
